refactor: Clean up project structure and add snap aim controller
- Add SnapAimController.h for enhanced targeting precision
- Update core infrastructure (performance_monitor, sync_manager, thread_safe_queue)
- Improve serial connection handling and input drivers
- Remove unused documentation and reference files
- Clean up temporary development files (optimization_fixes.cpp, serial_connection_fix.cpp)

+++ needaimbot/mouse/input_drivers/SerialConnection.cpp
@@ -398,9 +398,9 @@ bool SerialConnection::configurePort()
     const char* move_cmd = "m0,0\n";
     
     DWORD bytes_written;
-    WriteFile(serial_handle_, release_cmd, strlen(release_cmd), &bytes_written, NULL);
+    WriteFile(serial_handle_, release_cmd, static_cast<DWORD>(strlen(release_cmd)), &bytes_written, NULL);
     Sleep(10);
-    WriteFile(serial_handle_, move_cmd, strlen(move_cmd), &bytes_written, NULL);
+    WriteFile(serial_handle_, move_cmd, static_cast<DWORD>(strlen(move_cmd)), &bytes_written, NULL);
     Sleep(10);
     
     return true;
